refactored DnD code into a Draggable component
